+++ infra/resources/ced/_modules/portal_be_container_app/api_portal_be_apim.tf
@@ -46,7 +46,7 @@ resource "azurerm_api_management_api" "ced_portal_be_v1" {
 
   import {
     content_format = "openapi-link"
-    content_value  = "https://raw.githubusercontent.com/pagopa/io-growth/fe9e2d5ca502c0964749d9a6212bd3c179486557/apps/ced-portal-be/openapi/exposed/openapi.yaml"
+    content_value  = "https://raw.githubusercontent.com/pagopa/io-growth/d74999f4dde82e43f98cd2141bb61bcb5ef21c4e/apps/ced-portal-be/openapi/exposed/openapi.yaml"
   }
 }