+++ apps/ced-browser-be/CHANGELOG.md
@@ -1,5 +1,25 @@
 # ced-browser-be
 
+## 0.1.5
+
+### Patch Changes
+
+- 6f5c703: add profile info and make user info optional
+
+## 0.1.4
+
+### Patch Changes
+
+- 4536003: fix place full-text search: replace italian dictionary with simple for prefix matching
+- 836f060: add profile and opportunity api
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [7bbb05b]
+  - @pagopa/io-core-adapter-fims@0.0.7
+
 ## 0.1.2
 
 ### Patch Changes